+++ libraries/libfile/tar.h
@@ -7,8 +7,8 @@ struct TARBlock
     char name[100];
     char typeflag;
     char linkname[100];
-    u32 size;
+    size_t size;
     char *data;
-} ;
+};
 
 bool tar_read(void *tarfile, TARBlock *block, uint index);
